Extract updatePnc function into its own file
