+++ packages/core/phase3/lib/performOperations.ts
@@ -8,20 +8,13 @@ import type PncUpdateRequest from "../types/PncUpdateRequest"
 import type PncUpdateRequestGenerator from "../types/PncUpdateRequestGenerator"
 
 import { PncOperation } from "../../types/PncOperation"
-import generatePncUpdateExceptionFromMessage, {
-  isPncLockError
-} from "../exceptions/generatePncUpdateExceptionFromMessage"
 import PncUpdateRequestError from "../types/PncUpdateRequestError"
 import disposalUpdatedGenerator from "./pncUpdateRequestGenerators/disposalUpdatedGenerator"
 import { normalDisposalGenerator } from "./pncUpdateRequestGenerators/normalDisposalGenerator"
 import penaltyHearingGenerator from "./pncUpdateRequestGenerators/penaltyHearingGenerator"
 import { remandGenerator } from "./pncUpdateRequestGenerators/remandGenerator"
 import sentenceDeferredGenerator from "./pncUpdateRequestGenerators/sentenceDeferredGenerator"
-
-export const MAXIMUM_PNC_LOCK_ERROR_RETRIES = 3
-const DELAY_FOR_PNC_LOCK_ERROR_RETRY = parseInt(process.env.DELAY_FOR_PNC_LOCK_ERROR_RETRY ?? "10000")
-
-const delayForPncLockErrorRetry = () => new Promise((resolve) => setTimeout(resolve, DELAY_FOR_PNC_LOCK_ERROR_RETRY))
+import updatePnc from "./updatePnc"
 
 const pncUpdateRequestGenerator: { [T in PncOperation]: PncUpdateRequestGenerator<T> } = {
   [PncOperation.DISPOSAL_UPDATED]: disposalUpdatedGenerator,
@@ -42,32 +35,6 @@ const generatePncUpdateRequest = <T extends PncOperation>(
   }
 
   return pncUpdateRequest
-}
-
-const updatePnc = async (
-  pncUpdateDataset: PncUpdateDataset,
-  pncUpdateRequest: PncUpdateRequest,
-  pncGateway: PncGatewayInterface
-) => {
-  const correlationId = pncUpdateDataset.AnnotatedHearingOutcome.HearingOutcome.Hearing.SourceReference.UniqueID
-
-  for (let pncLockErrorRetries = 0; pncLockErrorRetries < MAXIMUM_PNC_LOCK_ERROR_RETRIES; pncLockErrorRetries++) {
-    const pncUpdateResult = await pncGateway.update(pncUpdateRequest, correlationId)
-
-    if (isError(pncUpdateResult)) {
-      const pncExceptions = pncUpdateResult.messages.map(generatePncUpdateExceptionFromMessage)
-
-      if (pncExceptions.some(isPncLockError) && pncLockErrorRetries !== MAXIMUM_PNC_LOCK_ERROR_RETRIES - 1) {
-        await delayForPncLockErrorRetry()
-      } else {
-        pncUpdateDataset.Exceptions.push(...pncExceptions)
-
-        return pncUpdateResult
-      }
-    } else {
-      return pncUpdateResult
-    }
-  }
 }
 
 const performOperations = async (
